Update OptionPriceModels/QLOptionPriceModel to support changing default values
This gives consumers more flexibility in creating the right model. It also employs a
'try not to suck' mentality. The previous implementation would happily return all zero
values even though it could provide an estimate. It was returning zeros because the
option pricing model used didn't match the option contract's style (American/European).
Now, even if there's a mismatch, we'll attempt to construt the correct exercise model
to at least provide a value to the caller instead of silently failing and returning
zeroes.